chore: split function_2 test
