+++ packages/amplify-e2e-tests/src/__tests__/function_2c.test.ts
@@ -1,0 +1,72 @@
+import {
+  addApiWithoutSchema,
+  updateApiSchema,
+  addAuthWithDefault,
+  addFunction,
+  addS3StorageWithSettings,
+  addSimpleDDB,
+  AddStorageSettings,
+  amplifyPush,
+  amplifyPushAuth,
+  amplifyPushForce,
+  createNewProjectDir,
+  deleteProject,
+  deleteProjectDir,
+  getFunctionSrcNode,
+  getProjectMeta,
+  initJSProjectWithProfile,
+  invokeFunction,
+  overrideFunctionSrcNode,
+  updateFunction,
+  addAuthWithGroupsAndAdminAPI,
+  getFunction,
+  loadFunctionTestFile,
+  createRandomName,
+  generateRandomShortId,
+} from '@aws-amplify/amplify-e2e-core';
+import _ from 'lodash';
+
+describe('nodejs', () => {
+  describe('amplify add function with additional permissions', () => {
+    let projRoot: string;
+    beforeEach(async () => {
+      projRoot = await createNewProjectDir('fn-with-perm');
+    });
+
+    afterEach(async () => {
+      await deleteProject(projRoot);
+      deleteProjectDir(projRoot);
+    });
+
+    it('@model-backed lambda function should generate envvars TODOTABLE_NAME, TODOTABLE_ARN, GRAPHQLAPIIDOUTPUT', async () => {
+      await initJSProjectWithProfile(projRoot, {
+        name: 'modelbackedlambda',
+      });
+      await addApiWithoutSchema(projRoot, { transformerVersion: 1 });
+      await updateApiSchema(projRoot, 'modelbackedlambda', 'simple_model.graphql');
+
+      const fnName = `integtestfn${generateRandomShortId()}`;
+
+      await addFunction(
+        projRoot,
+        {
+          name: fnName,
+          functionTemplate: 'Hello World',
+          additionalPermissions: {
+            permissions: ['storage'],
+            choices: ['api', 'storage'],
+            resources: ['Todo:@model(appsync)'],
+            resourceChoices: ['Todo:@model(appsync)'],
+            operations: ['read'],
+          },
+        },
+        'nodejs',
+      );
+
+      const lambdaSource = getFunctionSrcNode(projRoot, fnName);
+      expect(lambdaSource.includes('TODOTABLE_NAME')).toBeTruthy();
+      expect(lambdaSource.includes('TODOTABLE_ARN')).toBeTruthy();
+      expect(lambdaSource.includes('GRAPHQLAPIIDOUTPUT')).toBeTruthy();
+    });
+  });
+});
